# assig4_golang-main/services/contact/internal/delivery/http/swagger/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "contact service on clean architecture",
        "title": "slurm contact service on clean architecture",
        "contact": {
            "name": "API Support",
            "email": "kolyadkons@gmail.com"
        },
        "license": {
            "name": "kolyadkons"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/groups/": {
            "get": {
                "description": "Метод позволяет получить список групп.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Метод позволяет получить список групп.",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Количество записей",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Смещение при получении записей",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "name",
                        "description": "Сортировка по полю",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/group.GroupList"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "404 Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Метод позволяет создать группу контактов.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Метод позволяет создать группу контактов.",
                "parameters": [
                    {
                        "description": "Данные по группе",
                        "name": "group",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/group.ShortGroup"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/group.GroupResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "404 Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/groups/{id}": {
            "get": {
                "description": "Метод позволяет получить данные по группе.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Метод позволяет получить данные по группе.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор группы контактов",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/group.GroupResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "404 Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Метод позволяет обновить данные группы.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Метод позволяет обновить данные группы.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор группы",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные по группе",
                        "name": "group",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/group.ShortGroup"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/group.GroupResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "404 Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Метод позволяет удалить группу.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Метод позволяет удалить группу.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор группы",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "404 Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "group.GroupList": {
            "type": "object",
            "properties": {
                "limit": {
                    "description": "Количество записей",
                    "type": "integer",
                    "default": 10,
                    "minimum": 0,
                    "example": 10
                },
                "list": {
                    "description": "Список групп",
                    "type": "array",
                    "minItems": 0,
                    "items": {
                        "$ref": "#/definitions/group.GroupResponse"
                    }
                },
                "offset": {
                    "description": "Смещение при получении записей",
                    "type": "integer",
                    "default": 0,
                    "minimum": 0,
                    "example": 20
                },
                "total": {
                    "description": "Всего",
                    "type": "integer",
                    "default": 0,
                    "minimum": 0,
                    "example": 10
                }
            }
        },
        "group.GroupResponse": {
            "type": "object",
            "required": [
                "createdAt",
                "id",
                "modifiedAt",
                "name"
            ],
            "properties": {
                "contactsAmount": {
                    "description": "Кол-во контактов в группе",
                    "type": "integer",
                    "default": 10,
                    "minimum": 0
                },
                "createdAt": {
                    "description": "Дата создания группы",
                    "type": "string"
                },
                "description": {
                    "description": "Описание",
                    "type": "string",
                    "maxLength": 1000,
                    "example": "Описание группы"
                },
                "id": {
                    "description": "Идентификатор группы",
                    "type": "string",
                    "format": "uuid",
                    "example": "00000000-0000-0000-0000-000000000000"
                },
                "modifiedAt": {
                    "description": "Дата последнего изменения группы",
                    "type": "string"
                },
                "name": {
                    "description": "Название группы",
                    "type": "string",
                    "maxLength": 100,
                    "example": "Название группы"
                }
            }
        },
        "group.ShortGroup": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "description": "Описание",
                    "type": "string",
                    "maxLength": 1000,
                    "example": "Описание группы"
                },
                "name": {
                    "description": "Название группы",
                    "type": "string",
                    "maxLength": 100,
                    "example": "Название группы"
                }
            }
        },
        "http.ErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "info": {},
                "message": {
                    "type": "string"
                }
            }
        }
    }
}